requirements.txt
flask==3.0.0
opencv-python-headless==4.8.1
numpy==1.24.3
pillow==10.0.1
scikit-learn==1.7.2
gunicorn==21.2.0
python-dotenv==1.0.0